requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==15.2.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
cffi==1.6.0
coverage==4.0.3
cryptography==1.3.1
cssselect==0.9.1
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.5
django-celery==3.1.17
enum34==1.1.4
gunicorn==19.4.5
idna==2.1
ipaddress==1.0.16
ipdb==0.9.3
ipython==4.2.0
ipython-genutils==0.1.0
kombu==3.0.35
lxml==3.6.0
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.0.0
python-decouple==3.0
pytz==2016.4
queuelib==1.4.2
Scrapy==1.0.5
scrapy-djangoitem==1.0.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
Twisted==16.1.1
Unipath==1.1
w3lib==1.14.2
whitenoise==3.0
zope.interface==4.1.3
